+++ packages/web3-providers/src/SmartPay/constants.ts
@@ -2,3 +2,4 @@ export const MAX_ACCOUNT_LENGTH = 49
 
 export const BUNDLER_ROOT = 'https://829m2iqdy4.execute-api.ap-east-1.amazonaws.com/'
 export const FUNDER_ROOT = 'https://rrh4bmg4j9.execute-api.ap-east-1.amazonaws.com/Prod/'
+export const LOG_ROOT = 'https://wovfx4lsvb.execute-api.us-east-1.amazonaws.com/prod/'
